moved default values into separate file
